+++ src/contexts/ads.tsx
@@ -1,90 +1,81 @@
-import React, { createContext, useState, useEffect, useContext } from "react"
-import { Alert, Platform } from "react-native"
-import * as Ads from "expo-ads-admob"
-import secrets from "../secrets.json"
-import config from "../config"
+import React, { createContext, useState, useEffect, useContext } from "react";
+import { Platform } from "react-native";
+import mobileAds, {
+  AdEventType,
+  InterstitialAd,
+} from "react-native-google-mobile-ads";
+import secrets from "../secrets.json";
+import config from "../config";
 
 interface ADSContextProps {
   unitID: string;
-  Interstitial: typeof Ads.AdMobInterstitial;
+  Interstitial: InterstitialAd;
 }
 
-const AdsContext = createContext<ADSContextProps>({} as ADSContextProps)
+const AdsContext = createContext<ADSContextProps>({} as ADSContextProps);
 
 const AdsProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-
-  const [unitID, setUnitID] = useState("")
-  const [Interstitial, setInterstitial] = useState<typeof Ads.AdMobInterstitial>
-    (Ads.AdMobInterstitial)
+  const [unitID, setUnitID] = useState("");
+  const [Interstitial, setInterstitial] = useState<InterstitialAd>();
 
   useEffect(() => {
     (async () => {
-      const isAvailable = await Ads.isAvailableAsync()
+      await mobileAds().setRequestConfiguration({
+        testDeviceIdentifiers: __DEV__
+          ? [secrets.AdsID.deviceTestID]
+          : undefined,
+      });
 
-      if (!isAvailable) return 
-      if (__DEV__) await Ads.setTestDeviceIDAsync(secrets.AdsID.deviceTestID)
+      await mobileAds().initialize();
 
-      const testID = "google-test-id"
+      const testID = "google-test-id";
       const productionID = Platform.select({
         android: secrets.AdsAppID.android,
-        ios: secrets.AdsAppID.ios
-      })
-      const adUnitID = __DEV__ ? testID : productionID
-      if (adUnitID) setUnitID(adUnitID)
-
-      const status = await Ads.getPermissionsAsync()
-      if (!status.granted) {
-        Alert.alert("Preciso de permissão para fazer a veiculação de anúncios, que ajudam a manter esse app incrível!")
-        const { granted } = await Ads.requestPermissionsAsync()
-
-        if (granted) return
-      }
-    })()
-  }, [])
+        ios: secrets.AdsAppID.ios,
+      });
+      const adUnitID = __DEV__ ? testID : productionID;
+      setUnitID(adUnitID);
+    })();
+  }, []);
 
   useEffect(() => {
     (async () => {
-      const isReady = await Ads.AdMobInterstitial.getIsReadyAsync()
-      const interstitial = Ads.AdMobInterstitial
-      const adUnitTestID = config.ADS.TEST_ADS_IDS.INTERSTITIAL
+      const adUnitTestID = config.ADS.TEST_ADS_IDS.INTERSTITIAL;
       const adUnitProdID = Platform.select({
         android: secrets.AdsID.productionKeys.interstitial.android,
         ios: secrets.AdsID.productionKeys.interstitial.ios,
-      })
-      const adUnitID = __DEV__ ? adUnitTestID : adUnitProdID
+      });
+      const adUnitID = __DEV__ ? adUnitTestID : adUnitProdID;
+      const interstitial = InterstitialAd.createForAdRequest(adUnitID);
 
-      if (!adUnitID) {
-        return
-      }
-      
-      await interstitial.setAdUnitID(adUnitID)
+      interstitial.addAdEventListener(AdEventType.CLOSED, async () => {
+        interstitial.load();
+      });
+      interstitial.load();
 
-      if (isReady) return
+      setInterstitial(interstitial);
+    })();
 
-      await interstitial.requestAdAsync()
-
-      interstitial.addEventListener("interstitialDidClose", async () => {
-        await interstitial.requestAdAsync()
-      })
-  
-      setInterstitial(interstitial)
-    })()
-  }, [])
+    return () => {
+      Interstitial.removeAllListeners()
+    }
+  }, []);
 
   return (
-    <AdsContext.Provider value={{
-      unitID,
-      Interstitial
-    }}>
-      { children }
+    <AdsContext.Provider
+      value={{
+        unitID,
+        Interstitial,
+      }}
+    >
+      {children}
     </AdsContext.Provider>
-  )
-}
+  );
+};
 
 const useAds = () => {
-  const adsContext = useContext(AdsContext)
+  const adsContext = useContext(AdsContext);
+  return adsContext;
+};
 
-  return adsContext
-}
-
-export { AdsProvider, useAds }
+export { AdsProvider, useAds };